refactor(trace): split god files + dedupe field patterns (#5105, #5110)
