+++ src/core/runner/lab.rs
@@ -28,8 +28,8 @@ mod trace_fetch_refs;
 
 pub use super::lab_selection::LabRunnerSelectionSource;
 pub use offload::{
-    execute_lab_offload, LabOffloadCommand, LabOffloadOutcome, LabOffloadRequest,
-    LabOffloadSourcePathMode, LabOffloadWorkspaceModePolicy,
+    execute_lab_offload, LabLocalExecutionPolicy, LabOffloadCommand, LabOffloadOutcome,
+    LabOffloadRequest, LabOffloadSourcePathMode, LabOffloadWorkspaceModePolicy,
 };
 
 #[cfg(test)]
