+++ src/core/runners.rs
@@ -37,18 +37,18 @@ pub use super::runner::{
     prepare_lab_runner_capability, refresh_mirrored_daemon_evidence,
     reportable_artifact_evidence_path, resolve_default_lab_runner, run_reverse_worker,
     runner_artifact_store_token, runner_exec_failure_error, runner_job_log_snapshot, status,
-    statuses, sync_workspace, LabOffloadCommand, LabOffloadOutcome, LabOffloadRequest,
-    LabOffloadSourcePathMode, LabOffloadWorkspaceModePolicy, LabRunnerCapabilityContract,
-    LabRunnerGateDecision, LabRunnerGateMode, LabRunnerSelectionSource,
-    ManagedRunnerSourceSyncPlan, PreparedLabRunnerCapability, RemoteArtifactDownload,
-    ReverseRunnerConnectOptions, ReverseRunnerWorkerOptions, ReverseRunnerWorkerOutput, Runner,
-    RunnerCapabilityPreflight, RunnerConnectReport, RunnerDisconnectReport, RunnerExecDiagnostics,
-    RunnerExecMode, RunnerExecOptions, RunnerExecOutput, RunnerFailureKind, RunnerKind,
-    RunnerRequiredTool, RunnerResourceMetrics, RunnerSession, RunnerSessionRole,
-    RunnerSessionState, RunnerSpec, RunnerStaleDaemonWarning, RunnerStatusReport,
-    RunnerToolRegistry, RunnerToolSpec, RunnerTunnelMode, RunnerWorkspaceApplyOptions,
-    RunnerWorkspaceApplyOutput, RunnerWorkspaceApplyStatus, RunnerWorkspaceSyncMode,
-    RunnerWorkspaceSyncOptions, RunnerWorkspaceSyncOutput,
+    statuses, sync_workspace, LabLocalExecutionPolicy, LabOffloadCommand, LabOffloadOutcome,
+    LabOffloadRequest, LabOffloadSourcePathMode, LabOffloadWorkspaceModePolicy,
+    LabRunnerCapabilityContract, LabRunnerGateDecision, LabRunnerGateMode,
+    LabRunnerSelectionSource, ManagedRunnerSourceSyncPlan, PreparedLabRunnerCapability,
+    RemoteArtifactDownload, ReverseRunnerConnectOptions, ReverseRunnerWorkerOptions,
+    ReverseRunnerWorkerOutput, Runner, RunnerCapabilityPreflight, RunnerConnectReport,
+    RunnerDisconnectReport, RunnerExecDiagnostics, RunnerExecMode, RunnerExecOptions,
+    RunnerExecOutput, RunnerFailureKind, RunnerKind, RunnerRequiredTool, RunnerResourceMetrics,
+    RunnerSession, RunnerSessionRole, RunnerSessionState, RunnerSpec, RunnerStaleDaemonWarning,
+    RunnerStatusReport, RunnerToolRegistry, RunnerToolSpec, RunnerTunnelMode,
+    RunnerWorkspaceApplyOptions, RunnerWorkspaceApplyOutput, RunnerWorkspaceApplyStatus,
+    RunnerWorkspaceSyncMode, RunnerWorkspaceSyncOptions, RunnerWorkspaceSyncOutput,
 };
 
 // Registry CRUD entry points (re-exported at the root for ergonomics; also
